rename helper file and add some documentation
