+++ hokusai/lib/branch_detection.py
@@ -1,4 +1,7 @@
 from hokusai.lib.common import shout
+# This method is looking up the local git default branch (HEAD) and returning its name
+# For example if HEAD is "main" it will return "main"
+# If for any reason the lookup fails, it will return "main" as the default.
 
 def detect_branch():
   try:
